api/services/person_test.go
package services

import (
	"database/sql"
	"testing"

	"github.com/DATA-DOG/go-sqlmock"
	"github.com/stretchr/testify/assert"

	"github.com/jacob-tech-challenge/api/models"
)

func TestGetAllPeople(t *testing.T) {
	db, mock, err := sqlmock.New()
	if err != nil {
		t.Fatalf("an error '%s' was not expected when opening a stub database connection", err)
	}
	defer db.Close()

	tests := []struct {
		name           string
		inputName      string
		inputAge       int
		mockSetup      func(sqlmock.Sqlmock)
		expectedPeople []models.Person
		expectedError  bool
	}{
		{
			name:      "Success - No filters",
			inputName: "",
			inputAge:  0,
			mockSetup: func(mock sqlmock.Sqlmock) {
				rows := sqlmock.NewRows([]string{"id", "first_name", "last_name", "type", "age"}).
					AddRow(1, "John", "Doe", "student", 25).
					AddRow(2, "Jane", "Smith", "teacher", 30)
				mock.ExpectQuery(`SELECT \* FROM person WHERE \(\$1 = '' OR first_name = \$1\) AND \(\$2 = 0 OR age = \$2\)`).
					WithArgs("", 0).
					WillReturnRows(rows)

				courseRows1 := sqlmock.NewRows([]string{"id", "name"}).
					AddRow(1, "Math").
					AddRow(2, "Science")
				mock.ExpectQuery(`SELECT c\.id, c\.name FROM "course" c JOIN "person_course" pc ON c\.id = pc\.course_id WHERE pc\.person_id = \$1`).
					WithArgs(1).
					WillReturnRows(courseRows1)

				courseRows2 := sqlmock.NewRows([]string{"id", "name"}).
					AddRow(3, "History")
				mock.ExpectQuery(`SELECT c\.id, c\.name FROM "course" c JOIN "person_course" pc ON c\.id = pc\.course_id WHERE pc\.person_id = \$1`).
					WithArgs(2).
					WillReturnRows(courseRows2)
			},
			expectedPeople: []models.Person{
				{
					ID:        1,
					FirstName: "John",
					LastName:  "Doe",
					Type:      "student",
					Age:       25,
					Courses:   []int{1, 2},
				},
				{
					ID:        2,
					FirstName: "Jane",
					LastName:  "Smith",
					Type:      "teacher",
					Age:       30,
					Courses:   []int{3},
				},
			},
			expectedError: false,
		},
		{
			name:      "Success - Filter by name",
			inputName: "John",
			inputAge:  0,
			mockSetup: func(mock sqlmock.Sqlmock) {
				rows := sqlmock.NewRows([]string{"id", "first_name", "last_name", "type", "age"}).
					AddRow(1, "John", "Doe", "student", 25)
				mock.ExpectQuery(`SELECT \* FROM person WHERE \(\$1 = '' OR first_name = \$1\) AND \(\$2 = 0 OR age = \$2\)`).
					WithArgs("John", 0).
					WillReturnRows(rows)

				courseRows := sqlmock.NewRows([]string{"id", "name"}).
					AddRow(1, "Math").
					AddRow(2, "Science")
				mock.ExpectQuery(`SELECT c\.id, c\.name FROM "course" c JOIN "person_course" pc ON c\.id = pc\.course_id WHERE pc\.person_id = \$1`).
					WithArgs(1).
					WillReturnRows(courseRows)
			},
			expectedPeople: []models.Person{
				{
					ID:        1,
					FirstName: "John",
					LastName:  "Doe",
					Type:      "student",
					Age:       25,
					Courses:   []int{1, 2},
				},
			},
			expectedError: false,
		},
		{
			name:      "Success - Filter by age",
			inputName: "",
			inputAge:  30,
			mockSetup: func(mock sqlmock.Sqlmock) {
				rows := sqlmock.NewRows([]string{"id", "first_name", "last_name", "type", "age"}).
					AddRow(2, "Jane", "Smith", "teacher", 30)
				mock.ExpectQuery(`SELECT \* FROM person WHERE \(\$1 = '' OR first_name = \$1\) AND \(\$2 = 0 OR age = \$2\)`).
					WithArgs("", 30).
					WillReturnRows(rows)

				courseRows := sqlmock.NewRows([]string{"id", "name"}).
					AddRow(3, "History")
				mock.ExpectQuery(`SELECT c\.id, c\.name FROM "course" c JOIN "person_course" pc ON c\.id = pc\.course_id WHERE pc\.person_id = \$1`).
					WithArgs(2).
					WillReturnRows(courseRows)
			},
			expectedPeople: []models.Person{
				{
					ID:        2,
					FirstName: "Jane",
					LastName:  "Smith",
					Type:      "teacher",
					Age:       30,
					Courses:   []int{3},
				},
			},
			expectedError: false,
		},
		{
			name:      "Success - No results",
			inputName: "NonExistent",
			inputAge:  0,
			mockSetup: func(mock sqlmock.Sqlmock) {
				// Return empty result set
				rows := sqlmock.NewRows([]string{"id", "first_name", "last_name", "type", "age"})
				mock.ExpectQuery(`SELECT \* FROM person WHERE \(\$1 = '' OR first_name = \$1\) AND \(\$2 = 0 OR age = \$2\)`).
					WithArgs("NonExistent", 0).
					WillReturnRows(rows)
			},
			expectedPeople: nil, // Changed from nil to empty slice to match implementation
			expectedError: false,
		},
		{
			name:      "Error - Database query error",
			inputName: "",
			inputAge:  0,
			mockSetup: func(mock sqlmock.Sqlmock) {
				mock.ExpectQuery(`SELECT \* FROM person WHERE \(\$1 = '' OR first_name = \$1\) AND \(\$2 = 0 OR age = \$2\)`).
					WithArgs("", 0).
					WillReturnError(sql.ErrConnDone)
			},
			expectedPeople: []models.Person{},
			expectedError: true,
		},
		{
			name:      "Error - Course query error",
			inputName: "",
			inputAge:  0,
			mockSetup: func(mock sqlmock.Sqlmock) {
				rows := sqlmock.NewRows([]string{"id", "first_name", "last_name", "type", "age"}).
					AddRow(1, "John", "Doe", "student", 25)
				mock.ExpectQuery(`SELECT \* FROM person WHERE \(\$1 = '' OR first_name = \$1\) AND \(\$2 = 0 OR age = \$2\)`).
					WithArgs("", 0).
					WillReturnRows(rows)

				mock.ExpectQuery(`SELECT c\.id, c\.name FROM "course" c JOIN "person_course" pc ON c\.id = pc\.course_id WHERE pc\.person_id = \$1`).
					WithArgs(1).
					WillReturnError(sql.ErrConnDone)
			},
			expectedPeople: nil, // This case can return nil since it's an error case
			expectedError: true,
		},
	}

	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			// Setup mock expectations
			tt.mockSetup(mock)

			// Execute the function
			people, err := GetAllPeople(db, tt.inputName, tt.inputAge)

			// Check error expectations
			if tt.expectedError {
				assert.Error(t, err)
			} else {
				assert.NoError(t, err)
				assert.Equal(t, tt.expectedPeople, people)
			}

			// Ensure all expectations were met
			if err := mock.ExpectationsWereMet(); err != nil {
				t.Errorf("there were unfulfilled expectations: %s", err)
			}
		})
	}
}

func TestGetPersonByName(t *testing.T) {
	db, mock, err := sqlmock.New()
	if err != nil {
		t.Fatalf("an error '%s' was not expected when opening a stub database connection", err)
	}
	defer db.Close()

	tests := []struct {
		name          string
		inputName     string
		mockSetup     func(sqlmock.Sqlmock)
		expectedPerson models.Person
		expectedError  bool
	}{
		{
			name:      "Success - Person found with courses",
			inputName: "John",
			mockSetup: func(mock sqlmock.Sqlmock) {
				// Mock the person query
				personRows := sqlmock.NewRows([]string{"id", "first_name", "last_name", "type", "age"}).
					AddRow(1, "John", "Doe", "student", 25)
				mock.ExpectQuery("SELECT \\* FROM person WHERE first_name = \\$1").
					WithArgs("John").
					WillReturnRows(personRows)

				// Mock the courses join query
				courseRows := sqlmock.NewRows([]string{"id", "name"}).
					AddRow(1, "Math").
					AddRow(2, "Science").
					AddRow(3, "History")
				mock.ExpectQuery(`SELECT c\.id, c\.name FROM "course" c JOIN "person_course" pc ON c\.id = pc\.course_id WHERE pc\.person_id = \$1`).
					WithArgs(1).
					WillReturnRows(courseRows)
			},
			expectedPerson: models.Person{
				ID:        1,
				FirstName: "John",
				LastName:  "Doe",
				Type:      "student",
				Age:       25,
				Courses:   []int{1, 2, 3},
			},
			expectedError: false,
		},
		{
			name:      "Success - Person found with no courses",
			inputName: "Jane",
			mockSetup: func(mock sqlmock.Sqlmock) {
				// Mock the person query
				personRows := sqlmock.NewRows([]string{"id", "first_name", "last_name", "type", "age"}).
					AddRow(2, "Jane", "Smith", "student", 22)
				mock.ExpectQuery("SELECT \\* FROM person WHERE first_name = \\$1").
					WithArgs("Jane").
					WillReturnRows(personRows)

				// Mock empty courses result
				courseRows := sqlmock.NewRows([]string{"id", "name"})
				mock.ExpectQuery(`SELECT c\.id, c\.name FROM "course" c JOIN "person_course" pc ON c\.id = pc\.course_id WHERE pc\.person_id = \$1`).
					WithArgs(2).
					WillReturnRows(courseRows)
			},
			expectedPerson: models.Person{
				ID:        2,
				FirstName: "Jane",
				LastName:  "Smith",
				Type:      "student",
				Age:       22,
				Courses:   nil, // Changed from empty slice to nil to match actual behavior
			},
			expectedError: false,
		},
		{
			name:      "Not Found - Empty Result",
			inputName: "NonExistent",
			mockSetup: func(mock sqlmock.Sqlmock) {
				mock.ExpectQuery("SELECT \\* FROM person WHERE first_name = \\$1").
					WithArgs("NonExistent").
					WillReturnError(sql.ErrNoRows) // Changed to return ErrNoRows instead of empty result
			},
			expectedPerson: models.Person{},
			expectedError: false,
		},
		{
			name:      "Error - Database Error",
			inputName: "John",
			mockSetup: func(mock sqlmock.Sqlmock) {
				mock.ExpectQuery("SELECT \\* FROM person WHERE first_name = \\$1").
					WithArgs("John").
					WillReturnError(sql.ErrConnDone)
			},
			expectedPerson: models.Person{},
			expectedError: true,
		},
		{
			name:      "Error - Courses Query Failed",
			inputName: "John",
			mockSetup: func(mock sqlmock.Sqlmock) {
				// Mock the person query success
				personRows := sqlmock.NewRows([]string{"id", "first_name", "last_name", "type", "age"}).
					AddRow(1, "John", "Doe", "student", 25)
				mock.ExpectQuery("SELECT \\* FROM person WHERE first_name = \\$1").
					WithArgs("John").
					WillReturnRows(personRows)

				// Mock courses query error
				mock.ExpectQuery(`SELECT c\.id, c\.name FROM "course" c JOIN "person_course" pc ON c\.id = pc\.course_id WHERE pc\.person_id = \$1`).
					WithArgs(1).
					WillReturnError(sql.ErrConnDone)
			},
			expectedPerson: models.Person{},
			expectedError: true,
		},
	}

	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			// Setup mock expectations
			tt.mockSetup(mock)

			// Execute the function
			person, err := GetPersonByName(db, tt.inputName)

			// Check error expectations
			if tt.expectedError {
				assert.Error(t, err)
			} else {
				assert.NoError(t, err)
				assert.Equal(t, tt.expectedPerson, person)
			}

			// Ensure all expectations were met
			if err := mock.ExpectationsWereMet(); err != nil {
				t.Errorf("there were unfulfilled expectations: %s", err)
			}
		})
	}
}


func TestUpdatePersonByName(t *testing.T) {
	db, mock, err := sqlmock.New(sqlmock.QueryMatcherOption(sqlmock.QueryMatcherEqual))
	if err != nil {
		t.Fatalf("Failed to create mock: %v", err)
	}
	defer db.Close()

	t.Run("Successful Update", func(t *testing.T) {
		// Test data
		oldName := "John"
		updatedPerson := models.Person{
			FirstName: "Johnny",
			LastName:  "Doe",
			Type:     "Student",
			Age:      25,
			Courses:  []int{1, 2},
		}
		
		sqlmock.NewRows([]string{"id", "first_name", "last_name", "type", "age"}).
					AddRow(1, "John", "Doe", "student", 25)

		// 1. Mock the UPDATE query
		mock.ExpectExec(`UPDATE person SET first_name = $1, last_name = $2, type = $3, age = $4 WHERE first_name = $5`).
			WithArgs(updatedPerson.FirstName, updatedPerson.LastName, updatedPerson.Type, updatedPerson.Age, oldName).
			WillReturnResult(sqlmock.NewResult(0, 1))

		// 2. Mock the SELECT query in GetPersonByName (using the NEW name)
		personRows := sqlmock.NewRows([]string{"id", "first_name", "last_name", "type", "age"}).
			AddRow(1, updatedPerson.FirstName, updatedPerson.LastName, updatedPerson.Type, updatedPerson.Age)

		mock.ExpectQuery(`SELECT * FROM person WHERE first_name = $1`).
			WithArgs(updatedPerson.FirstName).  // Use the new name here
			WillReturnRows(personRows)

		// 3. Mock the courses query
		courseRows := sqlmock.NewRows([]string{"id", "name"}).
			AddRow(1, "Mathematics").
			AddRow(2, "Physics")

		mock.ExpectQuery(`SELECT c.id, c.name FROM "course" c JOIN "person_course" pc ON c.id = pc.course_id WHERE pc.person_id = $1`).
			WithArgs(1).
			WillReturnRows(courseRows)

		// Execute the update
		result, err := UpdatePersonByName(db, oldName, updatedPerson)

		// Debug logging
		if err != nil {
			t.Logf("Error occurred: %v", err)
		} else {
			t.Logf("Update successful. Result: %+v", result)
		}

        // Assertions
		assert.NoError(t, err)
		if err == nil {
			assert.Equal(t, 1, result.ID)
			assert.Equal(t, updatedPerson.FirstName, result.FirstName)
			assert.Equal(t, updatedPerson.LastName, result.LastName)
			assert.Equal(t, updatedPerson.Type, result.Type)
			assert.Equal(t, updatedPerson.Age, result.Age)
			assert.Equal(t, updatedPerson.Courses, result.Courses)
		}

		// Verify all expectations were met
		if err := mock.ExpectationsWereMet(); err != nil {
			t.Errorf("there were unfulfilled expectations: %s", err)
		}
	})

	t.Run("Update Error", func(t *testing.T) {
		oldName := "John"
		updatedPerson := models.Person{
			FirstName: "Johnny",
			LastName:  "Doe",
			Type:     "Student",
			Age:      25,
		}

		// Mock the UPDATE query to return an error
		mock.ExpectExec(`UPDATE person SET first_name = $1, last_name = $2, type = $3, age = $4 WHERE first_name = $5`).
			WithArgs(updatedPerson.FirstName, updatedPerson.LastName, updatedPerson.Type, updatedPerson.Age, oldName).
			WillReturnError(sql.ErrConnDone)

		// Execute the update
		_, err := UpdatePersonByName(db, oldName, updatedPerson)

		// Assertions
		assert.Error(t, err)
		assert.Equal(t, sql.ErrConnDone, err)

		if err := mock.ExpectationsWereMet(); err != nil {
			t.Errorf("there were unfulfilled expectations: %s", err)
		}
	})
}

func TestCreatePerson(t *testing.T) {
    db, mock, err := sqlmock.New(sqlmock.QueryMatcherOption(sqlmock.QueryMatcherEqual))
    if err != nil {
        t.Fatalf("Error creating mock database: %v", err)
    }
    defer db.Close()

    tests := []struct {
        name          string
        inputPerson   models.Person
        mockBehavior  func(mock sqlmock.Sqlmock)
        expectedError bool
    }{
        {
            name: "Success",
            inputPerson: models.Person{
                FirstName: "John",
                LastName:  "Doe",
                Type:     "Student",
                Age:      25,
                Courses:  []int{1, 2},
            },
            mockBehavior: func(mock sqlmock.Sqlmock) {
                // Expect transaction to begin
                mock.ExpectBegin()

                // Expect INSERT into person with RETURNING clause
                mock.ExpectQuery("INSERT INTO person (first_name, last_name, type, age) VALUES ($1, $2, $3, $4) RETURNING id").
                    WithArgs("John", "Doe", "Student", 25).
                    WillReturnRows(sqlmock.NewRows([]string{"id"}).AddRow(1))

                // Expect INSERT into person_course for each course
                mock.ExpectExec("INSERT INTO person_course (person_id, course_id) VALUES ($1, $2)").
                    WithArgs(1, 1).
                    WillReturnResult(sqlmock.NewResult(1, 1))

                mock.ExpectExec("INSERT INTO person_course (person_id, course_id) VALUES ($1, $2)").
                    WithArgs(1, 2).
                    WillReturnResult(sqlmock.NewResult(1, 1))

                // Expect transaction to commit
                mock.ExpectCommit()
            },
            expectedError: false,
        },
    }

    for _, tt := range tests {
        t.Run(tt.name, func(t *testing.T) {
            tt.mockBehavior(mock)

            _, err := CreatePerson(db, tt.inputPerson)

            if tt.expectedError {
                assert.Error(t, err)
            } else {
                assert.NoError(t, err)
            }

            assert.NoError(t, mock.ExpectationsWereMet())
        })
    }
}

func TestDeletePersonByName(t *testing.T) {
	db, mock, err := sqlmock.New()
	if err != nil {
		t.Fatalf("Error creating mock database: %v", err)
	}
	defer db.Close()

	tests := []struct {
		name          string
		inputName     string
		mockBehavior  func(mock sqlmock.Sqlmock, name string)
		expectedError bool
	}{
		{
			name:      "Success",
			inputName: "John",
			mockBehavior: func(mock sqlmock.Sqlmock, name string) {
				// Expect SELECT query for ID
				rows := sqlmock.NewRows([]string{"id"}).AddRow(1)
				mock.ExpectQuery("SELECT id FROM person").
					WithArgs(name).
					WillReturnRows(rows)

				// Expect DELETE from person_course
				mock.ExpectExec("DELETE FROM person_course").
					WithArgs(1).
					WillReturnResult(sqlmock.NewResult(1, 1))

				// Expect DELETE from person
				mock.ExpectExec("DELETE FROM person").
					WithArgs(name).
					WillReturnResult(sqlmock.NewResult(1, 1))
			},
			expectedError: false,
		},
		{
			name:      "Person Not Found",
			inputName: "NonExistent",
			mockBehavior: func(mock sqlmock.Sqlmock, name string) {
				mock.ExpectQuery("SELECT id FROM person").
					WithArgs(name).
					WillReturnError(sql.ErrNoRows)
			},
			expectedError: true,
		},
		{
			name:      "Error Deleting from person_course",
			inputName: "John",
			mockBehavior: func(mock sqlmock.Sqlmock, name string) {
				rows := sqlmock.NewRows([]string{"id"}).AddRow(1)
				mock.ExpectQuery("SELECT id FROM person").
					WithArgs(name).
					WillReturnRows(rows)

				mock.ExpectExec("DELETE FROM person_course").
					WithArgs(1).
					WillReturnError(sql.ErrConnDone)
			},
			expectedError: true,
		},
	}

	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			tt.mockBehavior(mock, tt.inputName)

			err := DeletePersonByName(db, tt.inputName)

			if tt.expectedError {
				assert.Error(t, err)
			} else {
				assert.NoError(t, err)
			}

			assert.NoError(t, mock.ExpectationsWereMet())
		})
	}
}